moving `roformer` models to a separate folder
